GH-83: Support Hyperlink (requires 69f47cc)
Requires new `x-article-link-active` mixin.

The reason 0003 is added, not 0002, is cuz extending branch added 0002:
- site/frontera/task/GH-83-callout-element--GH-321-resize-callout-image
(there, the 0003 migration has 0002 as a dependency)
(here, the 0003 migration has 0001 as a dependency)

When a proper main-extending branch is made, this should still happen:
- task/GH-83-callout-element--GH-321-resize-callout-image
- task/GH-321-resize-callout-image
(there, the 0003 migration will have ??? as a dependency)

Good luck, Wes.