Update settings.py
Forgot to reset this after local development.

Note: This seems like a good setting to move into the `settings_custom` file (in the deployments repo) as a per portal value (once the CMS adopts the `Core-portal` setup for organizing settings, custom and secrets). Plus we can add the local override for a `settings.dev` to ease local dev.