Task/fp 1285 cms critical dependency upgrade (#437)
* Upgraded Pillow dependency to 8.4.0 and cleaned up some formatting in the settings.py file.

* Bunped the base image version for the cms to python 3.9.10-buster in order to support required upgrade to Pillow 9.0.0, which had a critical security vulnerability.

* Dropping python base image version for Django

Because we use Django 2.2.16 which supports Python 3.5–3.8 (but not 3.9).

* Update settings.py

Forgot to reset this after local development.

Note: This seems like a good setting to move into the `settings_custom` file (in the deployments repo) as a per portal value (once the CMS adopts the `Core-portal` setup for organizing settings, custom and secrets). Plus we can add the local override for a `settings.dev` to ease local dev.